fix(web,ui-web): extract server auth session mapper and fix AppTopNav header border color
